Move multi select code to package
