+++ collect_app/src/main/res/layout/delete_blank_form_layout.xml
@@ -22,7 +22,7 @@
 
     <androidx.fragment.app.FragmentContainerView
         android:id="@+id/buttons"
-        android:name="org.odk.collect.androidshared.ui.MultiSelectControlsFragment"
+        android:name="org.odk.collect.androidshared.ui.multiselect.MultiSelectControlsFragment"
         android:layout_width="match_parent"
         android:layout_height="0dp"
         app:layout_constraintBottom_toBottomOf="parent"
